+++ superset/views/dashboard/api.py
@@ -24,7 +24,6 @@ from marshmallow import fields, post_load, pre_load, Schema, ValidationError
 from marshmallow.validate import Length
 from sqlalchemy.exc import SQLAlchemyError
 
-from superset import appbuilder
 from superset.exceptions import SupersetException
 from superset.models.dashboard import Dashboard
 from superset.utils import core as utils
@@ -357,6 +356,3 @@ class DashboardRestApi(DashboardMixin, BaseSupersetModelRestApi):
             return self.response(200, message="OK")
         except SQLAlchemyError as e:
             return self.response_422(message=str(e))
-
-
-appbuilder.add_api(DashboardRestApi)
