+++ superset/views/database/api.py
@@ -18,7 +18,6 @@ from flask_appbuilder import ModelRestApi
 from flask_appbuilder.models.sqla.interface import SQLAInterface
 
 import superset.models.core as models
-from superset import appbuilder
 
 from .mixins import DatabaseFilter, DatabaseMixin
 from .validators import sqlalchemy_uri_validator
@@ -58,6 +57,3 @@ class DatabaseRestApi(DatabaseMixin, ModelRestApi):
     # Removes the local limit for the page size
     max_page_size = -1
     validators_columns = {"sqlalchemy_uri": sqlalchemy_uri_validator}
-
-
-appbuilder.add_api(DatabaseRestApi)
