+++ superset/views/log/api.py
@@ -18,7 +18,6 @@ from flask_appbuilder import ModelRestApi
 from flask_appbuilder.models.sqla.interface import SQLAInterface
 
 import superset.models.core as models
-from superset import app, appbuilder
 
 from . import LogMixin
 
@@ -39,10 +38,3 @@ class LogRestApi(LogMixin, ModelRestApi):
     allow_browser_login = True
     list_columns = ("user.username", "action", "dttm")
     show_columns = list_columns
-
-
-if (
-    not app.config["FAB_ADD_SECURITY_VIEWS"] is False
-    or app.config["SUPERSET_LOG_VIEW"] is False
-):
-    appbuilder.add_api(LogRestApi)
