+++ superset/views/schedules.py
@@ -23,10 +23,10 @@ from flask import flash, g
 from flask_appbuilder import expose
 from flask_appbuilder.models.sqla.interface import SQLAInterface
 from flask_appbuilder.security.decorators import has_access
-from flask_babel import gettext as __, lazy_gettext as _
+from flask_babel import lazy_gettext as _
 from wtforms import BooleanField, StringField
 
-from superset import app, appbuilder, db, security_manager
+from superset import db, security_manager
 from superset.exceptions import SupersetException
 from superset.models.dashboard import Dashboard
 from superset.models.schedules import (
@@ -271,29 +271,3 @@ class SliceEmailScheduleView(EmailScheduleView):  # pylint: disable=too-many-anc
         if item.slice is None:
             raise SupersetException("Slice is mandatory")
         super(SliceEmailScheduleView, self).pre_add(item)
-
-
-def _register_schedule_menus():
-    appbuilder.add_separator("Manage")
-
-    appbuilder.add_view(
-        DashboardEmailScheduleView,
-        "Dashboard Email Schedules",
-        label=__("Dashboard Emails"),
-        category="Manage",
-        category_label=__("Manage"),
-        icon="fa-search",
-    )
-
-    appbuilder.add_view(
-        SliceEmailScheduleView,
-        "Chart Emails",
-        label=__("Chart Email Schedules"),
-        category="Manage",
-        category_label=__("Manage"),
-        icon="fa-search",
-    )
-
-
-if app.config["ENABLE_SCHEDULED_EMAIL_REPORTS"]:
-    _register_schedule_menus()
